+++ lib/datura/to_es/xml_to_es.rb
@@ -51,6 +51,7 @@ class XmlToEs
   end
 
   def get_id
+    # no extension
     @filename
   end
 
@@ -64,6 +65,23 @@ class XmlToEs
 
   # see CommonXml module for methods replicated from common.xsl
 
+  # get_elements
+  #   returns a nodeset for the requested xpaths
+  #   typically this would be used in the event that refinable results are
+  #     needed (for example: creators with id and role) than can be obtained
+  #     from get_list or get_text
+  def get_elements(*xpaths, xml: nil)
+    doc = xml || @xml
+
+    # if an array was passed through instead of a string or list of strings,
+    # it is now nested in an array and we need to excavate it
+    if xpaths.length == 1 && xpaths.first.class == Array
+      xpaths = xpaths.first
+    end
+
+    doc.xpath(*xpaths) || []
+  end
+
   # get the value of one of the xpaths listed at the top
   # Note: if the xpath returns multiple values they will be squished together
   # TODO should we make it so that this can optionally look for more than one
@@ -72,9 +90,10 @@ class XmlToEs
   # get_list
   #   can pass it a string xpath or array of xpaths
   # returns an array with the html value in xpath
-  def get_list(xpaths, keep_tags=false, xml=nil)
-    xpath_array = xpaths.class == Array ? xpaths : [xpaths]
-    get_xpaths(xpath_array, keep_tags, xml)
+  def get_list(xpaths, keep_tags: false, xml: nil, sort: false)
+    xpath_array = Array(xpaths)
+    list = get_xpaths(xpath_array, keep_tags: keep_tags, xml: xml)
+    sort ? list.sort : list
   end
 
   # get_text
@@ -82,12 +101,10 @@ class XmlToEs
   #   can optionally set a delimiter, otherwise ;
   # returns a STRING
   # if you want a multivalued result, please refer to get_list
-  def get_text(xpaths, keep_tags=false, xml=nil, delimiter=";")
+  def get_text(xpaths, keep_tags: false, xml: nil, delimiter: ";", sort: false)
     # ensure all xpaths are an array before beginning
-    xpath_array = xpaths.class == Array ? xpaths : [xpaths]
-    list = get_xpaths(xpath_array, keep_tags, xml)
-    sorted = list.sort
-    sorted.join("#{delimiter} ")
+    list = get_list(xpaths, keep_tags: keep_tags, xml: xml, sort: sort)
+    list.join("#{delimiter} ")
   end
 
   # Note: Recommend that collection team do NOT use this method directly
@@ -95,28 +112,29 @@ class XmlToEs
   # keep_tags true will convert tags like <hi> to <em>
   #   use this wisely, as it causes performance issues
   # keep_tags false removes ALL tags from selected xpath
-  def get_xpaths(xpaths, keep_tags=false, xml=nil)
+  def get_xpaths(xpaths, keep_tags: false, xml: nil)
     doc = xml || @xml
     list = []
-    xpaths.each do |xpath|
-      contents = doc.xpath(xpath)
-      contents.each do |content|
-        text = ""
-        if keep_tags
-          converted = CommonXml.convert_tags(content)
-          text = converted.inner_html
-        else
-          # note: not using content.text because
-          # some tags should be converted to (), [], etc for display
-          text = CommonXml.to_display_text(content)
-        end
-        # remove whitespace of all kinds from the text
-        text = Datura::Helpers.normalize_space(text)
-        if text.length > 0
-          list << text
-        end
+
+    contents = get_elements(xpaths, xml: doc)
+    contents.each do |content|
+      text = ""
+
+      if keep_tags
+        converted = CommonXml.convert_tags(content)
+        text = converted.inner_html
+      else
+        # note: not using content.text because
+        # some tags should be converted to (), [], etc for display
+        text = CommonXml.to_display_text(content)
+      end
+      # remove whitespace of all kinds from the text
+      text = Datura::Helpers.normalize_space(text)
+      if text.length > 0
+        list << text
       end
     end
+
     list.uniq
   end